Desktop: Markdown editor: Support opening same-note links to anchors
Refactors jumpToHash to use the same logic as openLink
